建構以 configure 與 hadle 為主的 CLI 應用
