+++ crates/faq_crawler/Cargo.toml
@@ -4,17 +4,5 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-anyhow = "1.0.86"
 reqwest = { version = "0.12.5", features = ["blocking"] }
 scraper = "0.19.0"
-sea-orm = { version = "0.12.15", features = ["sqlx-postgres", "runtime-tokio-rustls", "macros"] }
-tokio = { version = "1.38.0", features = ["full"] }
-tracing = "0.1.40"
-tracing-subscriber = "0.3.18"
-uuid = { version = "1.9.1", features = ["v4"] }
-migration = {path = "../migration"}
-clap = "4.5.8"
-dotenv = "0.15.0"
-serde = { version = "1.0.203", features = ["derive"] }
-serde_json = "1.0.120"
-config = "0.14.0"
