--- crates/faq_crawler/src/commands/mod.rs
@@ -1,13 +1,0 @@
-use clap::{ArgMatches, Command};
-
-mod moi;
-
-pub fn configuration(command: Command) -> Command {
-    command.subcommand(moi::configure())
-}
-
-pub fn handler(matches: &ArgMatches) -> anyhow::Result<()> {
-    moi::handle(matches)?;
-
-    Ok(())
-}
