--- crates/faq_crawler/src/lib.rs
@@ -1,5 +1,0 @@
-#[macro_use]
-extern crate tracing;
-
-pub mod commands;
-pub mod settings;
