--- crates/faq_crawler/src/settings.rs
@@ -1,36 +1,0 @@
-use config::{Config, Environment, File};
-use serde::Deserialize;
-
-#[derive(Debug, Deserialize, Default)]
-pub struct Database {
-    pub url: String,
-}
-
-#[derive(Debug, Deserialize, Default)]
-pub struct Logging {
-    pub log_level: Option<String>,
-}
-
-#[derive(Debug, Deserialize, Default)]
-pub struct Settings {
-    #[serde(default)]
-    pub database: Database,
-    #[serde(default)]
-    pub logging: Logging,
-}
-
-impl Settings {
-    pub fn new(location: &str, env_prefix: &str) -> anyhow::Result<Self> {
-        let s = Config::builder()
-            .add_source(File::with_name(location))
-            .add_source(
-                Environment::with_prefix(env_prefix)
-                    .separator("__")
-                    .prefix_separator("__"),
-            )
-            .build()?;
-
-        let settings = s.try_deserialize()?;
-        Ok(settings)
-    }
-}
